moved create_gc_tasks from raft to coordinator
