+++ crates/indexify_proto/src/indexify_coordinator.rs
@@ -197,6 +197,15 @@ pub struct RemoveIngestionServerRequest {
 pub struct RemoveIngestionServerResponse {}
 #[allow(clippy::derive_partial_eq_without_eq)]
 #[derive(Clone, PartialEq, ::prost::Message)]
+pub struct CreateGcTasksRequest {
+    #[prost(string, tag = "1")]
+    pub content_id: ::prost::alloc::string::String,
+}
+#[allow(clippy::derive_partial_eq_without_eq)]
+#[derive(Clone, PartialEq, ::prost::Message)]
+pub struct CreateGcTasksResponse {}
+#[allow(clippy::derive_partial_eq_without_eq)]
+#[derive(Clone, PartialEq, ::prost::Message)]
 pub struct CoordinatorCommand {
     #[prost(message, optional, tag = "1")]
     pub gc_task: ::core::option::Option<GcTask>,
@@ -1112,6 +1121,36 @@ pub mod coordinator_service_client {
                 );
             self.inner.unary(req, path, codec).await
         }
+        pub async fn create_gc_tasks(
+            &mut self,
+            request: impl tonic::IntoRequest<super::CreateGcTasksRequest>,
+        ) -> std::result::Result<
+            tonic::Response<super::CreateGcTasksResponse>,
+            tonic::Status,
+        > {
+            self.inner
+                .ready()
+                .await
+                .map_err(|e| {
+                    tonic::Status::new(
+                        tonic::Code::Unknown,
+                        format!("Service was not ready: {}", e.into()),
+                    )
+                })?;
+            let codec = tonic::codec::ProstCodec::default();
+            let path = http::uri::PathAndQuery::from_static(
+                "/indexify_coordinator.CoordinatorService/CreateGCTasks",
+            );
+            let mut req = request.into_request();
+            req.extensions_mut()
+                .insert(
+                    GrpcMethod::new(
+                        "indexify_coordinator.CoordinatorService",
+                        "CreateGCTasks",
+                    ),
+                );
+            self.inner.unary(req, path, codec).await
+        }
         pub async fn gc_tasks_stream(
             &mut self,
             request: impl tonic::IntoStreamingRequest<
@@ -1602,6 +1641,13 @@ pub mod coordinator_service_server {
             request: tonic::Request<super::RemoveIngestionServerRequest>,
         ) -> std::result::Result<
             tonic::Response<super::RemoveIngestionServerResponse>,
+            tonic::Status,
+        >;
+        async fn create_gc_tasks(
+            &self,
+            request: tonic::Request<super::CreateGcTasksRequest>,
+        ) -> std::result::Result<
+            tonic::Response<super::CreateGcTasksResponse>,
             tonic::Status,
         >;
         /// Server streaming response type for the GCTasksStream method.
@@ -2397,6 +2443,53 @@ pub mod coordinator_service_server {
                     let fut = async move {
                         let inner = inner.0;
                         let method = RemoveIngestionServerSvc(inner);
+                        let codec = tonic::codec::ProstCodec::default();
+                        let mut grpc = tonic::server::Grpc::new(codec)
+                            .apply_compression_config(
+                                accept_compression_encodings,
+                                send_compression_encodings,
+                            )
+                            .apply_max_message_size_config(
+                                max_decoding_message_size,
+                                max_encoding_message_size,
+                            );
+                        let res = grpc.unary(method, req).await;
+                        Ok(res)
+                    };
+                    Box::pin(fut)
+                }
+                "/indexify_coordinator.CoordinatorService/CreateGCTasks" => {
+                    #[allow(non_camel_case_types)]
+                    struct CreateGCTasksSvc<T: CoordinatorService>(pub Arc<T>);
+                    impl<
+                        T: CoordinatorService,
+                    > tonic::server::UnaryService<super::CreateGcTasksRequest>
+                    for CreateGCTasksSvc<T> {
+                        type Response = super::CreateGcTasksResponse;
+                        type Future = BoxFuture<
+                            tonic::Response<Self::Response>,
+                            tonic::Status,
+                        >;
+                        fn call(
+                            &mut self,
+                            request: tonic::Request<super::CreateGcTasksRequest>,
+                        ) -> Self::Future {
+                            let inner = Arc::clone(&self.0);
+                            let fut = async move {
+                                <T as CoordinatorService>::create_gc_tasks(&inner, request)
+                                    .await
+                            };
+                            Box::pin(fut)
+                        }
+                    }
+                    let accept_compression_encodings = self.accept_compression_encodings;
+                    let send_compression_encodings = self.send_compression_encodings;
+                    let max_decoding_message_size = self.max_decoding_message_size;
+                    let max_encoding_message_size = self.max_encoding_message_size;
+                    let inner = self.inner.clone();
+                    let fut = async move {
+                        let inner = inner.0;
+                        let method = CreateGCTasksSvc(inner);
                         let codec = tonic::codec::ProstCodec::default();
                         let mut grpc = tonic::server::Grpc::new(codec)
                             .apply_compression_config(
